03_uns_graphdb/requirements.txt
neo4j>=4.4.1
paho-mqtt>=1.6.1
dynaconf
logger
